auth controller code refactoring
