+++ controller/auth/index.js
@@ -1,0 +1,16 @@
+const {controllerWrapper} = require("../../helpers");
+const register = require("./register");
+const login = require("./login");
+const getCurrent = require("./getCurrent");
+const logout = require("./logout");
+const updateSubscription = require("./updateSubscription");
+const updateAvatar = require("./updateAvatar");
+
+module.exports = {
+  register: controllerWrapper(register),
+  login: controllerWrapper(login),
+  getCurrent: controllerWrapper(getCurrent),
+  logout: controllerWrapper(logout),
+  updateSubscription: controllerWrapper(updateSubscription),
+  updateAvatar: controllerWrapper(updateAvatar),
+};
