+++ controller/auth/login.js
@@ -1,0 +1,40 @@
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+
+const {User} = require("../..//models/user");
+
+const {HttpError} = require("../..//helpers");
+
+const {SECRET_KEY} = process.env;
+
+const login = async (req, res) => {
+  const {email, password} = req.body;
+  const user = await User.findOne({email});
+
+  if (!user) {
+    throw HttpError(401, "Email or password is wrong");
+  }
+
+  const passwordCompare = await bcrypt.compare(password, user.password);
+
+  if (!passwordCompare) {
+    throw HttpError(401, "Email or password is wrong");
+  }
+
+  const token = jwt.sign({id: user._id}, SECRET_KEY, {expiresIn: "23h"});
+  await User.findByIdAndUpdate(user._id, {token});
+
+  res.json({
+    status: "OK",
+    code: 200,
+    data: {
+      token,
+      user: {
+        email: user.email,
+        subscription: user.subscription,
+      },
+    },
+  });
+};
+
+module.exports = login;
